shhhhhh

+++ manage/cmd/migration/scripts/25_add_webhook_type.sql
@@ -1,6 +1,6 @@
 ALTER TABLE webhooks
     ADD COLUMN type int NOT NULL DEFAULT 1,
-    ADD COLUMN source_id int DEFAULT NULL;
+    ADD COLUMN source_id bigint DEFAULT NULL;
 
 ALTER TABLE messages
     ADD COLUMN sticker_ids jsonb DEFAULT '[]';
